refactor(team): split lifecycle operation use cases
