+++ src/main/services/team/provisioning/TeamProvisioningMemberLifecycleServiceUseCases.ts
@@ -4,29 +4,17 @@ import {
 } from './TeamProvisioningOpenCodeMemberRestartSystemMessageUseCase';
 
 import type { AppendDirectProcessRuntimeEventUseCase } from './TeamProvisioningAppendDirectProcessRuntimeEventUseCase';
-import type { TeamProvisioningMemberLifecycleUseCasePorts } from './TeamProvisioningMemberLifecycle';
-import type { TeamProvisioningMemberLifecycleOperationRunner } from './TeamProvisioningMemberLifecycleOperationRunner';
 
 export interface TeamProvisioningMemberLifecycleServiceUseCasePorts {
   persistSentMessage(teamName: string, message: Record<string, unknown>): void;
   appendDirectProcessRuntimeEvent: AppendDirectProcessRuntimeEventUseCase;
-  operationRunner: Pick<
-    TeamProvisioningMemberLifecycleOperationRunner,
-    'runMemberLifecycleOperation'
-  >;
   nowIso(): string;
   randomUUID(): string;
 }
 
-export interface TeamProvisioningMemberLifecycleServiceUseCases extends Pick<
-  TeamProvisioningMemberLifecycleUseCasePorts,
-  | 'persistOpenCodeMemberRestartSystemMessage'
-  | 'appendDirectProcessRuntimeEvent'
-  | 'runMemberLifecycleOperation'
-> {
+export interface TeamProvisioningMemberLifecycleServiceUseCases {
   persistOpenCodeMemberRestartSystemMessage: PersistOpenCodeMemberRestartSystemMessageUseCase;
   appendDirectProcessRuntimeEvent: AppendDirectProcessRuntimeEventUseCase;
-  runMemberLifecycleOperation: TeamProvisioningMemberLifecycleOperationRunner['runMemberLifecycleOperation'];
 }
 
 export function createTeamProvisioningMemberLifecycleServiceUseCases(
@@ -40,7 +28,5 @@ export function createTeamProvisioningMemberLifecycleServiceUseCases(
         randomUUID: ports.randomUUID,
       }),
     appendDirectProcessRuntimeEvent: ports.appendDirectProcessRuntimeEvent,
-    runMemberLifecycleOperation: (teamName, memberName, kind, operation) =>
-      ports.operationRunner.runMemberLifecycleOperation(teamName, memberName, kind, operation),
   };
 }
