+++ src/main/services/team/provisioning/__tests__/TeamProvisioningMemberLifecycleServiceUseCases.test.ts
@@ -3,32 +3,11 @@ import { describe, expect, it } from 'vitest';
 import { createTeamProvisioningMemberLifecycleServiceUseCases } from '../TeamProvisioningMemberLifecycleServiceUseCases';
 
 import type { DirectProcessRuntimeEventInput } from '../TeamProvisioningAppendDirectProcessRuntimeEventUseCase';
-import type { TeamProvisioningMemberLifecycleOperationRunner } from '../TeamProvisioningMemberLifecycleOperationRunner';
-
-type RunMemberLifecycleOperationKind = Parameters<
-  TeamProvisioningMemberLifecycleOperationRunner['runMemberLifecycleOperation']
->[2];
 
 describe('TeamProvisioningMemberLifecycleServiceUseCases', () => {
   it('creates only the service-owned lifecycle use case ports', async () => {
     const sentMessages: Array<{ teamName: string; message: Record<string, unknown> }> = [];
     const runtimeEvents: DirectProcessRuntimeEventInput[] = [];
-    const operationCalls: string[] = [];
-    const operationRunner = {
-      marker: 'runner-bound',
-      async runMemberLifecycleOperation<T>(
-        this: { marker: string },
-        teamName: string,
-        memberName: string,
-        kind: RunMemberLifecycleOperationKind,
-        operation: () => Promise<T>
-      ): Promise<T> {
-        operationCalls.push(`${this.marker}:${teamName}:${memberName}:${kind}`);
-        return await operation();
-      },
-    } as Pick<TeamProvisioningMemberLifecycleOperationRunner, 'runMemberLifecycleOperation'> & {
-      marker: string;
-    };
 
     const useCases = createTeamProvisioningMemberLifecycleServiceUseCases({
       persistSentMessage: (teamName, message) => {
@@ -37,7 +16,6 @@ describe('TeamProvisioningMemberLifecycleServiceUseCases', () => {
       appendDirectProcessRuntimeEvent: async (input) => {
         runtimeEvents.push(input);
       },
-      operationRunner,
       nowIso: () => '2026-07-06T17:00:00.000Z',
       randomUUID: () => 'uuid-1',
     });
@@ -45,7 +23,6 @@ describe('TeamProvisioningMemberLifecycleServiceUseCases', () => {
     expect(Object.keys(useCases).sort()).toEqual([
       'appendDirectProcessRuntimeEvent',
       'persistOpenCodeMemberRestartSystemMessage',
-      'runMemberLifecycleOperation',
     ]);
 
     useCases.persistOpenCodeMemberRestartSystemMessage({
@@ -67,10 +44,6 @@ describe('TeamProvisioningMemberLifecycleServiceUseCases', () => {
       bootstrapRunId: 'bootstrap-1',
       source: 'test',
     });
-    await expect(
-      useCases.runMemberLifecycleOperation('team-a', 'Worker', 'manual_restart', async () => 'done')
-    ).resolves.toBe('done');
-
     expect(sentMessages).toHaveLength(1);
     expect(sentMessages[0]?.message).toMatchObject({
       from: 'Lead',
@@ -87,6 +60,5 @@ describe('TeamProvisioningMemberLifecycleServiceUseCases', () => {
         pid: 123,
       }),
     ]);
-    expect(operationCalls).toEqual(['runner-bound:team-a:Worker:manual_restart']);
   });
 });
